# example/petstore/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "petstore api",
        "title": "petstore api",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/pet": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "\"Update an existing pet by Id\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "pet"
                ],
                "summary": "\"Update an existing pet\"",
                "parameters": [
                    {
                        "description": "Update an existing pet by Id",
                        "name": "Pet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pet.PetDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/pet.PetDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "\"Add a new pet to the store\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "pet"
                ],
                "summary": "\"Add a new pet to the store\"",
                "parameters": [
                    {
                        "description": "Add a new pet to the store",
                        "name": "Pet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pet.PetDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/pet.PetDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/pet/findByStatus": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "\"Multiple status values can be provided with comma separated strings\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "pet"
                ],
                "summary": "\"Finds Pets by status\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status values that need to be considered for filter",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pet.PetDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/pet/findByTags": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "\"Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "pet"
                ],
                "summary": "\"Finds Pets by tags\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tags to filter by",
                        "name": "tags",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pet.PetDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid tag value"
                    }
                }
            }
        },
        "/pet/{petId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "\"Returns a single pet\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "pet"
                ],
                "summary": "\"Find pet by ID\"",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of pet to return",
                        "name": "petId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/pet.PetDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "\"default description\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "pet"
                ],
                "summary": "\"Updates a pet in the store with form data\"",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of pet that needs to be updated",
                        "name": "petId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of pet that needs to be updated",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status of pet that needs to be updated",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "\"delete a pet\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "pet"
                ],
                "summary": "\"Deletes a pet\"",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pet id to delete",
                        "name": "petId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid pet value"
                    }
                }
            }
        },
        "/pet/{petId}/uploadImage": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "\"default description\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "pet"
                ],
                "summary": "\"uploads an image\"",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of pet to update",
                        "name": "petId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Additional Metadata",
                        "name": "additionalMetadata",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/apiResponse.ApiResponseDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiResponse.ApiResponseDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "category.CategoryDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "pet.PetDTO": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/category.CategoryDTO"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "photoUrls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "description": "pet status in the store",
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tag.TagDTO"
                    }
                }
            }
        },
        "tag.TagDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}